+++ merchant-portal-lagom-app-api/src/main/scala/com/squareoneinsights/merchantportallagomapp/api/response/MerchantImpactDataResp.scala
@@ -16,7 +16,7 @@ object MerchantImpactDataResp {
   }
 }
 
-case class PaymentType(paymentAllow: Int, paymentInReview: Int, paymentBlock: Int)
+case class PaymentType(paymentAllow: Double, paymentInReview: Double, paymentBlock: Double)
 
 object PaymentType {
